make standalone functions member functions
