--- core/core-rt/BaseClient.cpp
@@ -1,57 +1,0 @@
-/*
- * Copyright 2019 The Nakama Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#include <future>
-
-#include "nakama-cpp/NClientInterface.h"
-#include "nakama-cpp/realtime/NWebsocketsFactory.h"
-#include "nakama-cpp/log/NLogger.h"
-#include "NRtClient.h"
-#include "BaseClient.h"
-
-#undef NMODULE_NAME
-#define NMODULE_NAME "Nakama::BaseClient"
-
-using namespace std;
-
-namespace Nakama {
-
-#if defined(BUILD_WEBSOCKET_WSLAY) && defined(BUILD_CURL_IO)
-NRtClientPtr BaseClient::createRtClient()
-{
-    return createRtClient(createDefaultWebsocket(_platformParams));
-}
-#endif
-
-NRtClientPtr BaseClient::createRtClient(NRtTransportPtr transport)
-{
-    RtClientParameters parameters;
-    parameters.host = _host;
-    parameters.port = _port;
-    parameters.ssl  = _ssl;
-    parameters.platformParams = _platformParams;
-
-    if (!transport)
-    {
-        NLOG_ERROR("No websockets transport passed. Please set transport.");
-        return nullptr;
-    }
-
-    NRtClientPtr client(new NRtClient(transport, parameters.host, parameters.port, parameters.ssl));
-    return client;
-}
-
-}
